pm view: share the npm field-path getter, drop the dead Expr path helpers, reject unknown dist-tags
Moves the field-path resolver into cli/npm_queryable.rs so bun pm pkg can
use the same grammar, deletes Expr::get_path_may_be_index/get_by_index
(pm view was their only caller), and treats a spec that is neither a
dist-tag nor a semver range as matching nothing instead of everything.
