+++ src/ast/expr.rs
@@ -366,130 +366,6 @@ impl Expr {
 
     // toJS alias deleted — `to_js` lives in `bun_js_parser_jsc::expr_jsc` extension trait.
 
-    /// Only use this for pretty-printing JSON. Do not use in transpiler.
-    ///
-    /// This does not handle edgecases like `-1` or stringifying arbitrary property lookups.
-    pub(crate) fn get_by_index(&self, index: u32, index_str: &[u8], bump: &Bump) -> Option<Expr> {
-        match &self.data {
-            Data::EArray(array) => {
-                if index >= array.items.len_u32() {
-                    return None;
-                }
-                Some(array.items.slice()[index as usize])
-            }
-            Data::EArrayJSON(array) => {
-                let items = array.get().items();
-                let item = items.get(index as usize)?;
-                Some(Expr::from_json_value(item, self.loc))
-            }
-            Data::EObjectJSON(object) => object
-                .get()
-                .properties()
-                .iter()
-                .find(|p| p.key.slice() == index_str)
-                .map(|p| Expr::from_json_value(&p.value, p.key_loc)),
-            Data::EObject(object) => {
-                for prop in object.properties.slice_const() {
-                    let Some(key) = &prop.key else { continue };
-                    match &key.data {
-                        Data::EString(str) => {
-                            if str.eql_bytes(index_str) {
-                                return prop.value;
-                            }
-                        }
-                        Data::ENumber(num) => {
-                            if num.to_u32() == index {
-                                return prop.value;
-                            }
-                        }
-                        _ => {}
-                    }
-                }
-                None
-            }
-            Data::EString(str) => {
-                let mut str = *str;
-                if str.len() > index as usize {
-                    let slice = str.slice(bump);
-                    // TODO: this is not correct since .length refers to UTF-16 code units and not UTF-8 bytes
-                    // However, since this is only used in the JSON prettifier for `bun pm view`, it's not a blocker for shipping.
-                    if slice.len() > index as usize {
-                        return Some(Expr::init(
-                            E::String {
-                                data: Str::new(&slice[index as usize..][..1]),
-                                ..Default::default()
-                            },
-                            self.loc,
-                        ));
-                    }
-                }
-                None
-            }
-            _ => None,
-        }
-    }
-
-    /// This supports lookups like:
-    /// - `foo`
-    /// - `foo.bar`
-    /// - `foo[123]`
-    /// - `foo[123].bar`
-    /// - `foo[123].bar[456]`
-    /// - `foo[123].bar[456].baz`
-    /// - `foo[123].bar[456].baz.qux` // etc.
-    ///
-    /// This is not intended for use by the transpiler, instead by pretty printing JSON.
-    // The arena is threaded explicitly because get_by_index allocates an
-    // E.String slice into &Bump.
-    pub fn get_path_may_be_index(&self, bump: &Bump, name: &[u8]) -> Option<Expr> {
-        if name.is_empty() {
-            return None;
-        }
-
-        if let Some(idx) = bun_core::strings::index_of_any(name, b"[.") {
-            match name[idx] {
-                b'[' => {
-                    let end_idx = bun_core::strings::index_of_char_usize(name, b']')? as usize;
-                    let mut base_expr = *self;
-                    if idx > 0 {
-                        let key = &name[..idx];
-                        base_expr = base_expr.get(key)?;
-                    }
-
-                    let index_str: &[u8] = &name[idx + 1..end_idx];
-                    // Path segments are bytes, not UTF-8.
-                    let index: u32 = bun_core::parse_unsigned(index_str, 10).ok()?;
-                    let rest: &[u8] = if name.len() > end_idx {
-                        &name[end_idx + 1..]
-                    } else {
-                        b""
-                    };
-                    let result = base_expr.get_by_index(index, index_str, bump)?;
-                    if !rest.is_empty() {
-                        return result.get_path_may_be_index(bump, rest);
-                    }
-                    return Some(result);
-                }
-                b'.' => {
-                    let key = &name[..idx];
-                    let sub_expr = self.get(key)?;
-                    let subpath: &[u8] = if name.len() > idx {
-                        &name[idx + 1..]
-                    } else {
-                        b""
-                    };
-                    if !subpath.is_empty() {
-                        return sub_expr.get_path_may_be_index(bump, subpath);
-                    }
-                    return Some(sub_expr);
-                }
-                _ => unreachable!(),
-            }
-        }
-
-        self.get(name)
-    }
-
     /// Don't use this if you care about performance.
     ///
     /// Sets the value of a property, creating it if it doesn't exist.
